modificando product order petition

+++ database/migrations/2025_08_08_143724_create_product_warehouses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained('products')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('warehouse_id')->constrained('warehouses')->onUpdate('cascade')->onDelete('cascade');
-            $table->integer('stock');
+            $table->decimal('stock', 10, 2);
             $table->timestamps();
         });
     }
